requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
filelock==3.0.12
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql-connector-python==8.0.11
numpy==1.21.1
platformdirs==2.0.2
protobuf==3.17.3
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
webdriver-manager==3.4.2
webencodings==0.5.1
xpath-py==0.1.2
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
filelock==3.0.12
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql-connector-python==8.0.11
numpy==1.21.1
Pillow==8.3.2
platformdirs==2.0.2
protobuf==3.17.3
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
webdriver-manager==3.4.2
webencodings==0.5.1
xpath-py==0.1.2
